Move smoke test scripts to scripts/smoke-tests/ subdirectory
Consolidate all 6 factory smoke tests into a dedicated subdirectory.
Update package.json script paths and fix relative imports accordingly.
