+++ packages/software-factory/package.json
@@ -9,13 +9,13 @@
     "boxel:search": "NODE_NO_WARNINGS=1 ts-node --transpileOnly scripts/boxel-search.ts",
     "boxel:session": "NODE_NO_WARNINGS=1 ts-node --transpileOnly scripts/boxel-session.ts",
     "cache:prepare": "NODE_NO_WARNINGS=1 ts-node --transpileOnly src/cli/cache-realm.ts",
-    "factory:agent-smoke": "NODE_NO_WARNINGS=1 ts-node --transpileOnly scripts/factory-agent-smoke.ts",
-    "factory:context-smoke": "NODE_NO_WARNINGS=1 ts-node --transpileOnly scripts/factory-context-smoke.ts",
+    "factory:agent-smoke": "NODE_NO_WARNINGS=1 ts-node --transpileOnly scripts/smoke-tests/factory-agent-smoke.ts",
+    "factory:context-smoke": "NODE_NO_WARNINGS=1 ts-node --transpileOnly scripts/smoke-tests/factory-context-smoke.ts",
     "factory:go": "NODE_NO_WARNINGS=1 ts-node --transpileOnly src/cli/factory-entrypoint.ts",
-    "factory:loop-smoke": "NODE_NO_WARNINGS=1 ts-node --transpileOnly scripts/factory-loop-smoke.ts",
-    "factory:prompt-smoke": "NODE_NO_WARNINGS=1 ts-node --transpileOnly scripts/factory-prompt-smoke.ts",
-    "factory:skill-smoke": "NODE_NO_WARNINGS=1 ts-node --transpileOnly scripts/factory-skill-smoke.ts",
-    "factory:tools-smoke": "NODE_NO_WARNINGS=1 ts-node --transpileOnly scripts/factory-tools-smoke.ts",
+    "factory:loop-smoke": "NODE_NO_WARNINGS=1 ts-node --transpileOnly scripts/smoke-tests/factory-loop-smoke.ts",
+    "factory:prompt-smoke": "NODE_NO_WARNINGS=1 ts-node --transpileOnly scripts/smoke-tests/factory-prompt-smoke.ts",
+    "factory:skill-smoke": "NODE_NO_WARNINGS=1 ts-node --transpileOnly scripts/smoke-tests/factory-skill-smoke.ts",
+    "factory:tools-smoke": "NODE_NO_WARNINGS=1 ts-node --transpileOnly scripts/smoke-tests/factory-tools-smoke.ts",
     "lint": "concurrently \"pnpm:lint:*(!fix)\" --names \"lint:\"",
     "lint:fix": "concurrently \"pnpm:lint:*:fix\" --names \"fix:\"",
     "lint:js": "eslint . --report-unused-disable-directives --cache",
