+++ packages/software-factory/scripts/smoke-tests/factory-loop-smoke.ts
@@ -17,9 +17,9 @@ import type {
   ProjectCard,
   TestResult,
   TicketCard,
-} from './lib/factory-agent';
+} from '../lib/factory-agent';
 
-import type { FactoryTool, ToolCallEntry } from './lib/factory-tool-builder';
+import type { FactoryTool, ToolCallEntry } from '../lib/factory-tool-builder';
 
 import {
   runFactoryLoop,
@@ -30,7 +30,7 @@ import {
   type FactoryLoopResult,
   type LoopAgent,
   type TestRunner,
-} from './lib/factory-loop';
+} from '../lib/factory-loop';
 
 // ---------------------------------------------------------------------------
 // Helpers
